Move getExtensionStatus/groupIdentitiesByChain into logic/extension
