+++ packages/bierzo-wallet/src/routes/login/index.tsx
@@ -7,10 +7,11 @@ import { Dispatch } from "redux";
 import { history } from "..";
 import BillboardMessage from "../../components/BillboardMessage";
 import { subscribeBalance } from "../../logic/balances";
+import { getExtensionStatus } from "../../logic/extension";
 import { drinkFaucetIfNeeded } from "../../logic/faucet";
 import { subscribeTransaction } from "../../logic/transactions";
 import { addBalancesAction, getBalances } from "../../store/balances";
-import { getExtensionStatus, setExtensionStateAction } from "../../store/extension";
+import { setExtensionStateAction } from "../../store/extension";
 import { addTickersAction, getTokens } from "../../store/tokens";
 import { addUsernamesAction, getUsernames } from "../../store/usernames/actions";
 import { BALANCE_ROUTE } from "../paths";
