+++ packages/bierzo-wallet/src/store/balances/index.unit.spec.ts
@@ -1,9 +1,10 @@
 import { parseGetIdentitiesResponse } from "../../communication/identities";
 import * as identities from "../../communication/identities";
 import { disconnect } from "../../logic/connection";
+import { getExtensionStatus } from "../../logic/extension";
 import { aNewStore } from "../../store";
 import { withChainsDescribe } from "../../utils/test/testExecutor";
-import { getExtensionStatus, setExtensionStateAction } from "../extension";
+import { setExtensionStateAction } from "../extension";
 import { addBalancesAction, getBalances } from "./actions";
 
 withChainsDescribe("Tokens reducer", () => {
