+++ packages/bierzo-wallet/src/store/usernames/index.unit.spec.ts
@@ -3,9 +3,10 @@ import { Address, ChainId } from "@iov/bcp";
 import { parseGetIdentitiesResponse } from "../../communication/identities";
 import * as identities from "../../communication/identities";
 import { disconnect } from "../../logic/connection";
+import { getExtensionStatus } from "../../logic/extension";
 import { aNewStore } from "../../store";
 import { withChainsDescribe } from "../../utils/test/testExecutor";
-import { getExtensionStatus, setExtensionStateAction } from "../extension";
+import { setExtensionStateAction } from "../extension";
 import { addUsernamesAction, getUsernames } from "./actions";
 import { BwUsername } from "./reducer";
